Cleanup waveform and etree thumbnail - start to use item-tile-image component instead
