+++ src/tiles/list/tile-list-compact.ts
@@ -3,10 +3,13 @@ import { ifDefined } from 'lit/directives/if-defined.js';
 import { customElement, property } from 'lit/decorators.js';
 import { SortParam } from '@internetarchive/search-service';
 import DOMPurify from 'dompurify';
+
 import { TileModel } from '../../models';
 import { formatCount, NumberFormat } from '../../utils/format-count';
 import { formatDate, DateFormat } from '../../utils/format-date';
 import { accountLabel } from './account-label';
+
+import '../item-tile-image';
 import '../mediatype-icon';
 
 @customElement('tile-list-compact')
@@ -66,13 +69,13 @@ export class TileListCompact extends LitElement {
       return nothing;
     }
     return html`
-      <item-image
+      <item-tile-image
         .model=${this.model}
         .baseImageUrl=${this.baseImageUrl}
         .isListTile=${true}
         .isCompactTile=${true}
       >
-      </item-image>
+      </item-tile-image>
     `;
   }
